Using a Moose Role for FFI:
[The MooseX::Role::FFI is provided here to make the review of both
easier to do at the same time. It will be released separately if
we were to approve this PR.]

The MooseX::Role::FFI allows to specify the subroutine data and the
library name - as well as the fallback - as subroutines.

It will then generate the corresponding callback for each sub, which
can be reached via a method.

The `after` allows us to have direct access to the FFI object *before*
the callbacks are generated. (Which I think is when we need it for
the mangler to work, but I'm not sure.)

What I like about it:

* Metadata, the rest is generated.
* Alien support built-in.
* We don't need to have class-level variables (though we could change
  it to it instead of subroutines.)
* We don't need to check that it has been instantiated (like we do
  in _get_version and $HAS_SUBS)

What I don't like about it:

* We need to make another method call to get the callback, depsite
  it being a definition of the class, not the instance.
* If you have a typo in the callback name, you cannot catch it,
  because it's a string to retrieve it.

I'm not sure the benefits outweigh the drawbacks. I was thinking of
possibly providing *another* interface and saying "pick yours."